Replace List.c with new C++ implementation
As long as the code was written in C, it was limited to storing
pointers, and we couldn't take advantage of inlining. Replace it
with a compatible C++ implementation which works in a similar manner,
combining a doubly linked list with a hash table for string-based
lookup.

The new implementation doesn't use any of the old OpenBOR list code.
Since anallyst claimed ownership of that code, rewriting from scratch
has the nice side benefit of clearing up any outstanding issues around
ownership/licensing.

This commit also removes the "#pragma pack(4)" from depends.h because
it is pure evil. Instead, we just set the packing on structures where we
need it, and use the compiler's default packing for everything else.