Implement garbage collection for persistent objects
It uses tri-color mark-and-sweep, and seems to work well. The mark phase is
designed to be run incrementally so we don't have to "stop the world"; the
sweep phase isn't incremental yet but can be made that way with some effort.

To test, run "test/cycle-gc.c": before GC, there should be 6 unfreed objects,
after, there should be only 3.