Rework ObjectHeap to fix objects being freed while still in use
If an object was created and made persistent, then the persistent
reference was removed, the object would be freed even if the original
temporary reference was still in use -- see test/reference-persistence2.c.
A similar issue occurred if an already-persistent value was read into a
temp register and the persistent reference was removed while the
temporary reference remained -- see test/reference-persistence3.c.

ObjectHeap has been reworked to fix this rather fundamental design flaw.
There is now only one "heap" for all objects, persistent and temporary.
The refcount of an object (or list) is incremented if it's read into a
temporary register and hasn't already been read into a temporary
register since temporary references were last cleared. Temporary
references are cleared and temporary objects are freed when
Interpreter::runFunction() finishes running a function, much like before.

The rework has the convenient side effect of considerably simplifying
the implementation in ObjectHeap.cpp. It also has the less desirable
side effect of a slight performance hit to the "get" instruction.