Replace HRESULT with a new result type CCResult.
HRESULT, S_OK, and E_FAIL are all names that the OpenBOR script engine
took from the Windows API for some reason. Not only is it strange to use
them in cross-platform code, it will cause problems if depends.h is ever
included at the same time as windows.h on a Windows platform. That
problem hasn't come up in ChronoScript, but it has happened in OpenBOR
before, and could easily happen again in Chrono Crash.

The SUCCEEDED() and FAILED() function macros, also from Windows, are
still intact for now but will be removed soon.