Pull back in the Mesa implementation of the HAS_TRIVIAL_DESTRUCTOR macro
2a2f9cac85f0e3e09821 replaced the original implementation from Mesa with
an implementation based on C++11's <type_traits> header to fix
compilation with clang++. This is ChronoScript's only use of a C++
header, and the bug with clang compilation was fixed in Mesa a while ago.
So pull in their fixed implementation, and use the C++11 header version
only as a fallback.