Bring back the INC and DEC opcodes for the ++ and -- operators.
These opcodes were in OpenBOR script, but were replaced in ChronoScript
with just add/sub 1. But it turns out there's a performance benefit to
using simpler instructions with fewer operands to adjust loop variables.

This probably helps benchmarks more than "real code", but it's a simple
enough change, and it won't hurt anything.